web: Rename MainLayout
